App cofig core good
