+++ src/main/java/com/bookverse/development/packapps/views/Structures.java
@@ -2,6 +2,7 @@ package com.bookverse.development.packapps.views;
 
 import com.bookverse.development.packapps.core.AppConfigCore;
 import com.bookverse.development.packapps.models.Resources;
+import com.bookverse.development.packapps.utils.Alerts;
 import java.awt.BorderLayout;
 import java.awt.FlowLayout;
 import java.awt.event.MouseEvent;
@@ -29,19 +30,23 @@ public class Structures extends JDialog implements MouseListener {
     JPanel panel = new JPanel(new FlowLayout());
     panel.setBorder(AppConfigCore.getBorder("Select structure"));
 
-    btnStack = resources.getLabel("  Stack  ", AppConfigCore.TEXT_COLOR, panel, AppConfigCore.MEDIUM);
+    btnStack = resources
+        .getLabel("  Stack  ", AppConfigCore.TEXT_COLOR, panel, AppConfigCore.MEDIUM);
     btnStack.setBorder(AppConfigCore.BORDER_BLUE);
     btnStack.addMouseListener(this);
 
-    btnQueue = resources.getLabel("  Queue  ", AppConfigCore.TEXT_COLOR, panel, AppConfigCore.MEDIUM);
+    btnQueue = resources
+        .getLabel("  Queue  ", AppConfigCore.TEXT_COLOR, panel, AppConfigCore.MEDIUM);
     btnQueue.setBorder(AppConfigCore.BORDER_BLUE);
     btnQueue.addMouseListener(this);
 
-    btnArrays = resources.getLabel("  Arrays  ", AppConfigCore.TEXT_COLOR, panel, AppConfigCore.MEDIUM);
+    btnArrays = resources
+        .getLabel("  Arrays  ", AppConfigCore.TEXT_COLOR, panel, AppConfigCore.MEDIUM);
     btnArrays.setBorder(AppConfigCore.BORDER_BLUE);
     btnArrays.addMouseListener(this);
 
-    btnExit = resources.getLabel("  Return  ", AppConfigCore.MAIN_COLOR, panel, AppConfigCore.MEDIUM);
+    btnExit = resources
+        .getLabel("  Return  ", AppConfigCore.MAIN_COLOR, panel, AppConfigCore.MEDIUM);
     btnExit.setBorder(AppConfigCore.BORDER_RED);
     btnExit.addMouseListener(this);
 
@@ -55,7 +60,7 @@ public class Structures extends JDialog implements MouseListener {
     setTitle("Structures of data");
     AppConfigCore.fadeIn(this);
     parent.setVisible(false);
-    AppConfigCore.instruccionesEstructuras();
+    Alerts.instruccionesEstructuras();
     setVisible(true);
   }
